@@ -1,11 +1,11 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 21
+    compileSdkVersion 33
     defaultConfig {
         applicationId 'de.tudarmstadt.informatik.secuso.phishedu2'
-        minSdkVersion 10
-        targetSdkVersion 19
+        minSdkVersion 19
+        targetSdkVersion 33
         versionCode 31
         versionName "1.4.0"
     }
@@ -20,9 +20,18 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-v4:21.0.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation project(':BaseGameUtils')
     implementation 'com.google.code.gson:gson:2.3'
     implementation files('libs/android-json-rpc-0.2.jar')
-    implementation 'com.android.support:appcompat-v7:21.0.2'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+
+    constraints {
+        implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version") {
+            because("kotlin-stdlib-jdk7 is now a part of kotlin-stdlib")
+        }
+        implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version") {
+            because("kotlin-stdlib-jdk8 is now a part of kotlin-stdlib")
+        }
+    }
 }
